Reworked link system
All roads (except the first one) have parents, which is the road it got global_goal generated from. When solidifying their links, roads now query their parent's links_e to find what it should link_s to. This kind of solves a lot of smaller problems with linking at once, like when a road snaps to the end of a road whose children haven't all generated. Before, those new children would only link_s to their parent. Also, it fixes the problem that prompted this change, which is that roads that got split with children not all generated would have their children not link to the split road's other half. Although the code for snap_to_cross is still not as pretty as I would like, this solution is actually fairly elegant over all. I was not expecting to get this working so fast.